dashboard/requirements.txt
streamlit>=1.39.0
pandas>=2.2.0
requests>=2.32.0
websocket-client>=1.8.0
